split into model, service and controller
