+++ backend/src/main/java/org/example/backend/model/SoundSource.java
@@ -1,0 +1,4 @@
+package org.example.backend.model;
+
+public record SoundSource(String classpath, boolean mp3) {
+}
